fix: lock-free audio for mac
