+++ cradle/mac/Cargo.template.toml
@@ -28,7 +28,8 @@ env_logger = "0.5"
 objc = "0.2"
 appkit = { git = "https://github.com/Pctg-x8/appkit-rs" }
 libc = "0.2"
-parking_lot = "0.11"
+parking_lot = "0.12"
 tracing = { version = "0.1.37", features = ["log"] }
 tracing-subscriber = { version = "0.3.17", features = ["fmt", "env-filter"] }
 async-std = "1.13.0"
+thiserror = "2.0"
